converted timestamps to long

+++ src/main/java/com/chimpler/adnetwork/spout/RandomImpressionTupleSpout.java
@@ -1,20 +1,7 @@
 package com.chimpler.adnetwork.spout;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Random;
-import java.util.Set;
-
-import org.joda.time.DateTime;
-
-import storm.trident.operation.TridentCollector;
-import storm.trident.spout.IBatchSpout;
-import storm.trident.spout.ITridentSpout;
-import storm.trident.testing.FixedBatchSpout;
-import storm.trident.topology.TridentTopologyBuilder;
-
-import backtype.storm.Config;
-import backtype.storm.spout.ISpout;
 import backtype.storm.spout.SpoutOutputCollector;
 import backtype.storm.task.TopologyContext;
 import backtype.storm.topology.OutputFieldsDeclarer;
@@ -36,7 +23,7 @@ public class RandomImpressionTupleSpout extends BaseRichSpout {
 
 	@Override
 	public void nextTuple() {
-		int timestamp = (int)(System.currentTimeMillis() / 1000);
+		long timestamp = System.currentTimeMillis() / 1000;
 		int publisherId = random.nextInt(10);
 		int countryId = random.nextInt(10);
 		int cookieId = random.nextInt(10000);
